ui: ui 자잘 자잘한 수정
